refact: changed to deal with monetary values as integer

+++ src/domain/repository/taxes_repository.py
@@ -1,7 +1,9 @@
 from abc import ABC, abstractmethod
+
+
 class TaxesRepository(ABC):
     @abstractmethod
-    def get_by_id_and_type(self, id: str, type: str) -> float:
+    def get_by_id_and_type(self, id: str, type: str) -> int:
         """Search product's tax by ID and its type of tax
 
         Args:
@@ -9,6 +11,6 @@ class TaxesRepository(ABC):
             type (str): type of tax
 
         Returns:
-            float: Product ' tax'
+            int: Product ' tax'
         """
         raise NotImplementedError
