+++ src/infra/repository/memory/taxes_repository_memory.py
@@ -1,35 +1,11 @@
 from ....domain.repository.taxes_repository import TaxesRepository
 
-BOOK_DEFAULT_MONTH_TAX = {
-    'id': '1',
-    'type': 'default',
-    'value': 1/100
-}
-BOOK_NOVEMBER_TAX = {
-    'id': '1',
-    'type': 'november',
-    'value': 0.1/100
-}
-PFF2_MASK_DEFAULT_MONTH_TAX = {
-    'id': '2',
-    'type': 'default',
-    'value': 0.1/100
-}
-PFF2_MASK_NOVEMBER_TAX = {
-    'id': '2',
-    'type': 'november',
-    'value': 0
-}
-VACUUM_CLEANER_DEFAULT_MONTH_TAX = {
-    'id': '3',
-    'type': 'default',
-    'value': 10/100
-}
-VACUUM_CLEANER_NOVEMBER_TAX = {
-    'id': '3',
-    'type': 'november',
-    'value': 1/100
-}
+BOOK_DEFAULT_MONTH_TAX = {"id": "1", "type": "default", "value": 1 / 100}
+BOOK_NOVEMBER_TAX = {"id": "1", "type": "november", "value": 0.1 / 100}
+PFF2_MASK_DEFAULT_MONTH_TAX = {"id": "2", "type": "default", "value": 0.1 / 100}
+PFF2_MASK_NOVEMBER_TAX = {"id": "2", "type": "november", "value": 0}
+VACUUM_CLEANER_DEFAULT_MONTH_TAX = {"id": "3", "type": "default", "value": 10 / 100}
+VACUUM_CLEANER_NOVEMBER_TAX = {"id": "3", "type": "november", "value": 1 / 100}
 
 
 class TaxesRepositoryMemory(TaxesRepository):
@@ -42,7 +18,7 @@ class TaxesRepositoryMemory(TaxesRepository):
             VACUUM_CLEANER_DEFAULT_MONTH_TAX,
             VACUUM_CLEANER_NOVEMBER_TAX,
         ]
-    
+
     def get_by_id_and_type(self, id: str, type: str) -> float:
         """Search product's tax by ID and its type of tax
 
@@ -51,9 +27,9 @@ class TaxesRepositoryMemory(TaxesRepository):
             type (str): type of tax
 
         Returns:
-            float: Product ' tax'
+            int: Product ' tax'
         """
         for tax in self.__table:
-            if tax['id'] == id and tax['type'] == type:
-                return tax['value']
+            if tax["id"] == id and tax["type"] == type:
+                return tax["value"]
         return 0
